outsourcing context and state into a separate provider component
