+++ onlineShop/src/App.jsx
@@ -1,80 +1,12 @@
-import { useState } from 'react';
-
 import Header from './components/Header.jsx';
 import Shop from './components/Shop.jsx';
 import { DUMMY_PRODUCTS } from './dummy-products.js';
 import Product from './components/Product.jsx';
-import { CartContext } from './store/shoppingCart-context.jsx';
+import CartContextProvider from './store/shoppingCart-context.jsx';
 
 function App() {
-	const [shoppingCart, setShoppingCart] = useState({
-		items: [],
-	});
-
-	function handleAddItemToCart(id) {
-		setShoppingCart((prevShoppingCart) => {
-			const updatedItems = [...prevShoppingCart.items];
-
-			const existingCartItemIndex = updatedItems.findIndex(
-				(cartItem) => cartItem.id === id
-			);
-			const existingCartItem = updatedItems[existingCartItemIndex];
-
-			if (existingCartItem) {
-				const updatedItem = {
-					...existingCartItem,
-					quantity: existingCartItem.quantity + 1,
-				};
-				updatedItems[existingCartItemIndex] = updatedItem;
-			} else {
-				const product = DUMMY_PRODUCTS.find((product) => product.id === id);
-				updatedItems.push({
-					id: id,
-					name: product.title,
-					price: product.price,
-					quantity: 1,
-				});
-			}
-
-			return {
-				items: updatedItems,
-			};
-		});
-	}
-
-	function handleUpdateCartItemQuantity(productId, amount) {
-		setShoppingCart((prevShoppingCart) => {
-			const updatedItems = [...prevShoppingCart.items];
-			const updatedItemIndex = updatedItems.findIndex(
-				(item) => item.id === productId
-			);
-
-			const updatedItem = {
-				...updatedItems[updatedItemIndex],
-			};
-
-			updatedItem.quantity += amount;
-
-			if (updatedItem.quantity <= 0) {
-				updatedItems.splice(updatedItemIndex, 1);
-			} else {
-				updatedItems[updatedItemIndex] = updatedItem;
-			}
-
-			return {
-				items: updatedItems,
-			};
-		});
-	}
-
-	const contextValue = {
-		items: shoppingCart.items,
-		addItemToCart: handleAddItemToCart,
-		updatedItemQuantity: handleUpdateCartItemQuantity,
-	};
-
 	return (
-		<CartContext.Provider value={contextValue}>
+		<CartContextProvider>
 			<Header />
 			<Shop>
 				{DUMMY_PRODUCTS.map((product) => (
@@ -83,7 +15,7 @@ function App() {
 					</li>
 				))}
 			</Shop>
-		</CartContext.Provider>
+		</CartContextProvider>
 	);
 }
